+++ app/src/main/java/com/tetsoft/typego/data/language/LanguageList.kt
@@ -1,4 +1,4 @@
-package com.tetsoft.typego.data
+package com.tetsoft.typego.data.language
 
 import android.content.Context
 import com.tetsoft.typego.R
